improve redo log

+++ server/src/main/java/com/github/bdqfork/server/transaction/backup/AlwaysBackup.java
@@ -1,19 +1,14 @@
 package com.github.bdqfork.server.transaction.backup;
 
-import java.io.ByteArrayOutputStream;
 import java.io.DataOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.nio.charset.StandardCharsets;
-import java.util.*;
+import java.util.List;
 
-import com.github.bdqfork.core.exception.SerializeException;
 import com.github.bdqfork.server.transaction.RedoLog;
 import com.github.bdqfork.server.transaction.TransactionLog;
-
-import io.netty.util.Version;
 
 /**
  * @author bdq
@@ -37,7 +32,7 @@ public class AlwaysBackup extends AbstractBackupStrategy {
             List<RedoLog> redoLogs = transactionLog.getRedoLogs();
             for (RedoLog redoLog : redoLogs) {
                 byte[] data = getSerializer().serialize(redoLog);
-                dataOutputStream.writeInt(HEAD);
+                dataOutputStream.writeByte(HEAD);
                 dataOutputStream.writeByte(VERSION);
                 dataOutputStream.writeInt(data.length);
                 dataOutputStream.write(data);
